Make posts page the homepage, move old homepage to /old
The posts listing is now served at / instead of /posts.
The previous full-screen photo homepage is now at /old.
Updated canonical URLs and internal links accordingly.
