+++ pages/index.tsx
@@ -1,65 +1,79 @@
-import Image from 'next/image';
+import moment from 'moment';
+import { InferGetStaticPropsType } from 'next';
 import Link from 'next/link';
+import React from 'react';
 import Head from 'next/head';
-import KnightIcon from '../public/knight.svg';
-import portraitLandscape from '../public/portrait-landscape.jpg';
-import portraitSquare from '../public/portrait-square.jpg';
+import Colophon from '../components/Colophon';
+import { getPostDate, getPostSlug, getPostTitle, isPostPublic } from '../lib/utils';
+import { fetchCards } from '../lib/TrelloClient';
 
-type Link = {
-  href: string;
-  text: string;
-};
+export async function getStaticProps() {
+  const postsFromTrello = (await fetchCards())
+    .filter(item => isPostPublic(item))
+    .map(
+      (item): Post => ({
+        date: getPostDate(item) ?? null,
+        id: item.id,
+        slug: getPostSlug(item),
+        title: getPostTitle(item) ?? item.name,
+      }),
+    );
+  const staticPosts: Post[] = [
+    'reading',
+    'airbnb-tips',
+    'maine',
+    'now',
+    'saigon-roads',
+    'sapiens',
+    'timeline',
+    'favorite-albums',
+  ].map(slug => {
+    const meta: Record<string, string> = require(`./${slug}.mdx`).meta;
+    return {
+      id: slug,
+      slug,
+      date: meta.lastUpdated || meta.postDate || null,
+      title: meta.title,
+    };
+  });
+  return {
+    props: {
+      posts: [...postsFromTrello, ...staticPosts].sort((a, b) => {
+        const aDate = new Date(a.date);
+        const bDate = new Date(b.date);
+        if (aDate && bDate) {
+          return bDate.getTime() - aDate.getTime();
+        }
+        return 0;
+      }),
+    },
+    revalidate: 1,
+  };
+}
 
-const links: Link[] = [
-  { text: 'Posts', href: '/posts' },
-];
-
-export default function IndexPage() {
+export default function Posts({ posts }: InferGetStaticPropsType<typeof getStaticProps>) {
   return (
     <>
       <Head>
         <title>benknight.me</title>
-        <meta name="theme-color" content="#000" />
         <link rel="canonical" href="https://benknight.me/" />
       </Head>
-      <div className="fixed inset-0 flex flex-col items-center justify-center font-medium landscape:font-light text-white antialiased text-opacity-40">
-        <div className="absolute top-0 left-0 w-full h-full">
-          <Image
-            alt=""
-            className="portrait:hidden filter brightness-110"
-            layout="fill"
-            objectFit="cover"
-            placeholder="blur"
-            quality={90}
-            src={portraitLandscape}
-          />
-          <Image
-            alt=""
-            className="landscape:hidden filter brightness-110"
-            layout="fill"
-            objectFit="cover"
-            placeholder="blur"
-            quality={90}
-            sizes="100vh"
-            src={portraitSquare}
-          />
-        </div>
-        <ul className="relative uppercase text-sm leading-9 tracking-[0.5em] text-center">
-          {links.map(link => (
-            <li className="mb-3" key={link.href}>
-              <Link href={link.href}>
-                <a
-                  className="block hover:text-white hover:text-opacity-60 ease p-2"
-                  rel="author">
-                  {link.text}
-                </a>
-              </Link>
-            </li>
-          ))}
-          <li className="mt-8" title="It’s a knight, get it?">
-            <KnightIcon className="relative left-[-0.25rem] fill-current block w-4 h-4 mx-auto hover:text-white hover:text-opacity-60 transition-colors duration-500 ease" />
-          </li>
-        </ul>
+      <Colophon />
+      <div className="px-3 py-6 max-w-xl m-auto prose dark:prose-invert">
+        {posts.map(post => (
+          <div className="mb-4" key={post.id}>
+            <Link href={`/${post.slug}`}>
+              <a className="block p-4 group no-underline text-center text-black/60 dark:text-white/60">
+                <h3 className="inline m-0 underline underline-offset-2 text-xl text-inherit">
+                  <b>{post.title}</b>
+                </h3>
+                <small className="block font-light no-underline">
+                  {post.date && moment(post.date).format('LL')}
+                </small>
+              </a>
+            </Link>
+          </div>
+        ))}
       </div>
     </>
   );
